Move Base to data validation module
